+++ src/types/api/Token.ts
@@ -1,4 +1,4 @@
-export type Token = {
+export interface Token {
     access_token: string,
     expires_in: number,
     refresh_token: string,
@@ -6,4 +6,4 @@ export type Token = {
     token_type: string,
     user_id: string,
     expireDate: string,
-};
+}