+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 All notable changes to this project will be documented in this file.
 
 
+## [2.1.0-beta.4] 2019-03-21
+### Merge branch 'beta' of github.com
+- fabulator/fitbit-api-handler into beta
+
+### Changed
+- Polishing BREAKING
+
+## [2.1.0-beta.3] 2019-03-21
+### Changed
+- Update packages
+
+## [2.1.0-beta.2] 2019-01-15
+### Changed
+- Update fitnes models library
+
+## [2.1.0-beta.1] 2018-12-22
+### Changed
+- Use fitness models dependency for Activity model.
+
 ## [2.0.0] - 2018-11-17
 ### Changed
 - Migrate to typescript